back-end/requirements.txt
Django>=3.0,<4.0
djangorestframework==3.13.1
psycopg2-binary>=2.8
parsel==1.6.0
requests==2.24.0
pyngrok==5.2.1